+++ fractales/Julia.cpp
@@ -1,7 +1,7 @@
 #include <complex>
 #include "Julia.h"
 
-int Julia::compute_value(std::complex<double> z) {
+double Julia::compute_value(std::complex<double> z) {
     int i = 0;
     std::complex<double> c = -0.4 + 0.6i;
     while (std::norm(std::pow(z, 2) + c) < 4 && i < maxiter) {
