+++ fractales/MartinMandelbrot.cpp
@@ -27,7 +27,7 @@ void MartinMandelbrot::compute_seed(std::complex<double> s) {
     }
 }
 
-int MartinMandelbrot::compute_value(std::complex<double> z) {
+double MartinMandelbrot::compute_value(std::complex<double> z) {
     std::complex<double> delta = z - X;
     if (std::norm(delta) != 0 && std::norm(std::pow(delta, 3)) * 1000 > std::norm(std::pow(delta,2))) {
         compute_seed(z);
